[SharedCache] Optimize parsing of slide info
Slide info is parsed and applied on the main thread, below the
`SharedCache` constructor, when a shared cache is opened. Time spent
applying the slide is time that the main thread is blocked.

These changes eliminate some unnecessary overhead so what work remains
is dominated by kernel work (paging in data, copying pages when the
first modification is made).

The changes are:
1. Read slide pointers via `ReadULong` rather than the variable-length
   `Read` method. The compiler isn't able to eliminate the call to
   `memcpy` in the variable-length `Read` function, and the function
   call overhead is noticeable given the small size of the read and
   number of times it is called.
2. Remove the file member from `MappingInfo`. Slide info is applied for
   a single file at a time so there's no reason to track the file it
   belongs to. This removes unnecesssary reference counting on the
   `std::shared_ptr` that holds the `MMappedFileAccessor`.