[SharedCache] Allow the user to select the original shared cache file on disk
This gives the user the ability to take a database with no original file path continue processing entries with the user supplied base file.

This also will set the original file path for future loads, we might want to make that optional. And the behavior with remote projects for that might be undesired.