[ObjC] Fix handling of relative method selectors with MSVC
The order that the operands to `+` are evaluated in is unspecified. Clang
happens to evaluate them left to right and gives the expected answer.
MSVC picks the opposite order and so the value it computes is off by 4.

This was resulting in missing method names when arm64 binaries
containing Objective-C are analyzed on Windows.